+++ src/Levels.js
@@ -1,12 +1,12 @@
 export const levels = [
     { level:1 ,grid: 3, square: 3 },
     { level:2, grid: 3, square: 4 },
-    // { level:3, grid: 4, square: 4 },
-    // { level:4, grid: 4, square: 5 },
-    // { level:5, grid: 4, square: 6 },
-    // { level:6, grid: 5, square: 5 },
-    // { level:7,  grid: 5, square: 6 },
-    // { level:8,  grid: 5, square: 7 },
-    // { level:9,  grid: 6, square: 6 },
-    // { level:10,  grid: 7, square: 7 }
+    { level:3, grid: 4, square: 4 },
+    { level:4, grid: 4, square: 5 },
+    { level:5, grid: 4, square: 6 },
+    { level:6, grid: 5, square: 5 },
+    { level:7,  grid: 5, square: 6 },
+    { level:8,  grid: 5, square: 7 },
+    { level:9,  grid: 6, square: 6 },
+    { level:10,  grid: 7, square: 7 }
 ];
